Use global transactionContext
Currently, we pass Member in each call to chain's create/send transaction
methods. This patch introduces a method in chain to initialize the
transaction context once and use it every time a call to
createTransactionProposal, createDeploymentProposal, sendProposal or
sendTransaction is made. This simplifies the APIs for end user.

Also write test cases for validation, and fix some validation issues when
required arguments are not passed

Change-Id: I22d5ff13582c12a9e6b1074de7ca89be693546b1
Signed-off-by: Muhammad Altaf <muhammada@fast.au.fujitsu.com>