[FABJ-491] Default grpc/grpcs based on discovery peer
Currently grpcs is used for all discovered peers and
orderers. This is a bad assumption to make when the
network is not using TLS. If the discovery peer is
using TLS, then use grpcs, else use grpc. You cannot
have mixed TLS and non-TLS networks, so we don't need
to support discovery peer using TLS and discovered peer
not using TLS.

Change-Id: I94fa69bb3d93840f9cdb7dcf762313d47f86ad78
Signed-off-by: Simon Stone <sstone1@uk.ibm.com>