[FABJ-521] Allow arrays of PEM content in connection profile (#63)
The connection profile tlsCaCerts element has an optional "pem" property containing one or more PEM certificates. This change allows the "pem" property to be either a string or an array of strings. For arrays, all of the contained strings are concatenated to produce the same result as including multiple certificates in a single string.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>