Simplify pull request checks
Add a job to the end of the pull_request workflow that depends on the build jobs, and can be used as a single job to check pull request success. This avoids the need to change the branch protection PR checks when changes are made to the build process.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>
(cherry picked from commit 0ebb1cb46bbef227cdcb32fdcd00c9eaabcbe613)